Initialize stack trace in the error constructor.
This makes it possible for each "new Error()" to have a stack
right away, and not only after the error is used in a "throw" expression.

This requires creating an extra Exception object in most cases,
since the exception is often replaced when the Error is finally
thrown. It's my understanding that the big cost of exceptions is in
generating the stack trace, not the Exception object, and this
is still done lazily when the "stack" property is first accessed.